+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Censored sensitive connection properties (such as `password` and `tokenValue`) from the diagnostic details serialized into workflow and job submission command errors. [#TBD](https://github.com/zowe/zowe-cli/pull/TBD)
+
 ## `7.29.28`
 
 - BugFix: Updated the `zowe zos-files ds edit` command to store temp files in a dedicated subdirectory with owner-only permissions (`0o700`). [#2773](https://github.com/zowe/zowe-cli/pull/2773)